rename testing modules
